Use a overwritten runserver command as the solution instead
